secrets: add secret store
This adds a default builtin Secret Store as well as the implementation
to pass a Secret Store to the daemon upon initialization.

Signed-off-by: Evan Hazlett <ejhazlett@gmail.com>

secrets: use ramfs mount for secrets

This creates a ramfs and mounts in the container as read-only.  This
commit also updates the API to use the secret name instead of ID.

Signed-off-by: Evan Hazlett <ejhazlett@gmail.com>